+++ src/routes/layout.svelte
@@ -1,5 +1,7 @@
 <script lang="ts">
     import { app } from '@/lib/app/main';
+    import { isAndroidPermissionsGranted } from '@/lib/notification/repository';
+    import { getPlatform } from '@/lib/shared/platform';
     import { check } from '@tauri-apps/plugin-updater';
     import { onDestroy, onMount, type Snippet } from 'svelte';  
 
@@ -7,13 +9,21 @@
 
     let isUpdateAvailable: boolean = $state(false);
 
-    // On the page being mounted to the DOM, the authentication controller starts.
+    // On the page being mounted to the DOM, the application starts.
     onMount(async () => {
         app.start();
-        isUpdateAvailable = await check() !== null ? true : false ;
+
+        // This checks if update is available for the windows application. 
+        if (getPlatform() === "windows") {
+            isUpdateAvailable = await check() !== null ? true : false;
+        }
+
+        if (getPlatform() === "android") {
+            isAndroidPermissionsGranted()
+        }
     })
 
-    // On the page being unmounted from the DOM, the authentication controller stops.
+    // On the page being unmounted from the DOM, the application stops.
     onDestroy(() => {
         app.stop();
     })