+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # docket-todo
 
+## 2.1.0
+
+### Minor Changes
+
+- Implemented a notifications scheduler for the android and windows applications
+    - Notifications run based on the task's due date.
+    - Windows Notification Scheduler runs while the app is open. This will be improved upon in future updates.
+    - Android Notification Scheduler works even while the application is closed.
+
 ## 2.0.4
 
 ### Patch Changes